langchain_ver/requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.143
groq
streamlit
pymupdf
python-dotenv
faiss-cpu
sentence-transformers
pdfplumber
zstandard
pydantic>=2.0.0
reportlab
